requirements.txt
opencv-python
numpy
pyyaml
imgaug
pandas